requirements.txt
transformers
torch
datasets
faiss-cpu
flask
scikit-learn
 